+++ examples/plugin/host/src/Schemas/Routes/Login/Login.ts
@@ -1,0 +1,37 @@
+import {SchemaDefaultReturn} from 'figtree-schemas';
+import {ExtractSchemaResultType, Vts} from 'vts';
+
+export const SchemaTest = Vts.object({
+    test: Vts.or([
+        Vts.object({
+            test1: Vts.string(),
+            test2: Vts.string()
+        }),
+        Vts.string()
+    ])
+});
+
+export const SchemaIsLogin = SchemaDefaultReturn.extend({
+    status: Vts.boolean({description: 'Status is the user login'}),
+    more: Vts.optional(SchemaTest)
+});
+
+export type IsLogin = ExtractSchemaResultType<typeof SchemaIsLogin>;
+
+export const SchemaIsLoginParameter = Vts.object({
+    username: Vts.optional(Vts.string()),
+    userid: Vts.optional(Vts.string())
+});
+
+export type IsLoginParameter = ExtractSchemaResultType<typeof SchemaIsLoginParameter>;
+
+export const SchemaIsLoginParameterPath = Vts.object({
+    userid: Vts.string({description: 'Userid for request'})
+});
+
+export const SchemaLoginRequest = Vts.object({
+    username: Vts.string(),
+    password: Vts.string()
+});
+
+export type LoginRequest = ExtractSchemaResultType<typeof SchemaLoginRequest>;